Refactor recipient queries to share fragments
  https://eaflood.atlassian.net/browse/WATER-5559

  The licence holder and primary user SQL logic is essentially the same across the renewal invitations and abstraction alerts recipient queries. The only difference is the limiting conditions appended per use case (e.g. filtering by expiring licences or
  registered licences).

  To avoid duplication, the common SQL is extracted into shared DAL files:
  - `licence-holder-recipient-query.dal.js`
  - `primary-user-recipient-query.dal.js`

  Both `generate-renewal-recipients-query.service.js` and the new `generate-abstraction-alert-recipients-query.service.js` now compose their full queries by interpolating these shared fragments and appending their own `WHERE`/`JOIN` conditions.

  As part of this, the query-generation logic that was embedded in `fetch-abstraction-alert-recipients.service.js` is extracted into its own dedicated service, matching the pattern already established for renewal invitations.
